Change type of HTTPBody.Length.known from Int to Int64 (#79)
### Motivation

The associated value of enum case `HTTPBody.Length.known` was `Int`,
and, on watchOS, `Int = Int32`, which is not ideal for a content length.

### Modifications

- (API breaking) Change type of HTTPBody.Length.known from Int to Int64.

### Result

Can now express larger values for content type on 32-bit platforms.

### Test Plan

Unit tests pass.

### Related Issues

- Fixes https://github.com/apple/swift-openapi-generator/issues/354.

### Note

I have marked this PR as `semver/major` as it constitutes the first of
the PRs we make in the run up to 1.0.0-alpha.1 and the release notes
generator should therefore catch us from inadvertently cutting another
pre-1.0 patch release.

+++ Sources/OpenAPIRuntime/Multipart/MultipartFramesToRawPartsSequence.swift
@@ -77,7 +77,7 @@ extension HTTPBody {
         let stream = AsyncThrowingStream(unfolding: bodyClosure)
         let length: HTTPBody.Length
         if let contentLengthString = headerFields[.contentLength], let contentLength = Int(contentLengthString) {
-            length = .known(contentLength)
+            length = .known(Int64(contentLength))
         } else {
             length = .unknown
         }
